+++ src/init.c
@@ -3,32 +3,24 @@
 #include <R.h>
 #include <Rinternals.h>
 
-extern SEXP dbl_array4_to_packed_();
-extern SEXP dbl_array3_to_packed_();
-extern SEXP dbl_array1_to_packed_();
-
-extern SEXP packed_to_dbl_array4_();
-extern SEXP packed_to_dbl_array3_();
-extern SEXP packed_to_dbl_array1_();
+extern SEXP dbl_array_to_packed_();
+extern SEXP packed_to_dbl_array_();
 
 extern SEXP raw_to_packed_();
 extern SEXP packed_to_raw_();
-extern SEXP packed_set_dim_();
+
+extern SEXP packed_ptr_set_dim_();
 
 static const R_CallMethodDef CEntries[] = {
 
-  // name                       pointer                           Num args
-  {"dbl_array4_to_packed_" , (DL_FUNC) &dbl_array4_to_packed_ , 3},
-  {"dbl_array3_to_packed_"  , (DL_FUNC) &dbl_array3_to_packed_  , 3},
-  {"dbl_array1_to_packed_", (DL_FUNC) &dbl_array1_to_packed_, 3},
+  // name                       pointer                       Num args
+  {"dbl_array_to_packed_" , (DL_FUNC) &dbl_array_to_packed_ , 4},
+  {"packed_to_dbl_array_" , (DL_FUNC) &packed_to_dbl_array_ , 5},
 
-  {"packed_to_dbl_array4_" , (DL_FUNC) &packed_to_dbl_array4_ , 2},
-  {"packed_to_dbl_array3_"  , (DL_FUNC) &packed_to_dbl_array3_  , 2},
-  {"packed_to_dbl_array1_", (DL_FUNC) &packed_to_dbl_array1_, 2},
+  {"raw_to_packed_" , (DL_FUNC) &raw_to_packed_ , 3},
+  {"packed_to_raw_" , (DL_FUNC) &packed_to_raw_ , 1},
 
-  {"raw_to_packed_", (DL_FUNC) &raw_to_packed_, 3},
-  {"packed_to_raw_", (DL_FUNC) &packed_to_raw_, 1},
-  {"packed_set_dim_", (DL_FUNC) &packed_set_dim_, 3},
+  {"packed_ptr_set_dim_", (DL_FUNC) &packed_ptr_set_dim_, 3},
 
   {NULL , NULL, 0}
 };