+++ tests/testthat/test-basic.R
@@ -3,11 +3,11 @@
 
 test_that("conversion to/from grey matrix works", {
   mat <- matrix(seq(4*3*1), c(4, 3))/255
-  packed_ptr <- planar_to_packed(mat, format = 0L, maxval = 1)
+  packed_ptr <- planar_to_packed(mat, format = 0L, maxval = 1, as_ptr = TRUE)
   res  <- packed_to_planar(packed_ptr, format = 0L, nchannel = 1)
   expect_true(isTRUE(all.equal(mat, res)))
 
-  raw_vec <- packed_to_raw(packed_ptr)
+  packed_raw <- packed_to_raw(packed_ptr)
 
   raw_ref <- as.raw(c(
     1,  1,  1, 255,
@@ -24,12 +24,12 @@ test_that("conversion to/from grey matrix works", {
     12, 12, 12, 255
   ))
 
-  expect_identical(raw_vec, raw_ref)
+  expect_identical(packed_raw, raw_ref)
 
-  a2 <- raw_to_packed(raw_vec, width = 3L, height = 4L)
+  a2 <- raw_to_packed(packed_raw, width = 3L, height = 4L)
   r2 <- packed_to_raw(a2)
 
-  expect_identical(raw_vec, r2)
+  expect_identical(packed_raw, r2)
 
 })
 
@@ -37,12 +37,12 @@ test_that("conversion to/from grey matrix works", {
 
 test_that("conversion to/from rgb array works", {
   arr <- array(seq(4*3*3), c(4, 3, 3))/255
-  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1)
+  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1, as_ptr = TRUE)
   res  <- packed_to_planar(packed_ptr, format = 0L, nchannel = 3)
   expect_true(isTRUE(all.equal(arr, res)))
 
 
-  raw_vec <- packed_to_raw(packed_ptr)
+  packed_raw <- packed_to_raw(packed_ptr)
 
   raw_ref <- as.raw(c(
     0x19, 0x0d, 0x01, 0xff, 0x1d, 0x11, 0x05, 0xff, 0x21,
@@ -52,24 +52,24 @@ test_that("conversion to/from rgb array works", {
     0x08, 0xff, 0x24, 0x18, 0x0c, 0xff
   ))
 
-  expect_identical(raw_vec, raw_ref)
+  expect_identical(packed_raw, raw_ref)
 
-  a2 <- raw_to_packed(raw_vec, width = 3L, height = 4L)
+  a2 <- raw_to_packed(packed_raw, width = 3L, height = 4L)
   r2 <- packed_to_raw(a2)
 
-  expect_identical(raw_vec, r2)
+  expect_identical(packed_raw, r2)
 })
 
 
 
 test_that("conversion to/from rgba array works", {
   arr <- array(seq(4*3*4), c(4, 3, 4))/255
-  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1)
+  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1, as_ptr = TRUE)
   res  <- packed_to_planar(packed_ptr, format = 0L, nchannel = 4)
   expect_true(isTRUE(all.equal(arr, res)))
 
 
-  raw_vec <- packed_to_raw(packed_ptr)
+  packed_raw <- packed_to_raw(packed_ptr)
 
   raw_ref <- as.raw(c(
     0x19, 0x0d, 0x01, 0x25, 0x1d, 0x11, 0x05, 0x29, 0x21,
@@ -79,24 +79,24 @@ test_that("conversion to/from rgba array works", {
     0x08, 0x2c, 0x24, 0x18, 0x0c, 0x30
   ))
 
-  expect_identical(raw_vec, raw_ref)
+  expect_identical(packed_raw, raw_ref)
 
-  a2 <- raw_to_packed(raw_vec, width = 3L, height = 4L)
+  a2 <- raw_to_packed(packed_raw, width = 3L, height = 4L)
   r2 <- packed_to_raw(a2)
 
-  expect_identical(raw_vec, r2)
+  expect_identical(packed_raw, r2)
 
 
 
-  expect_error(raw_to_packed(raw_vec, width = 3L, height = 5L), "Length")
+  expect_error(raw_to_packed(packed_raw, width = 3L, height = 5L), "Length")
 })
 
 
 
 test_that("set_dim ok", {
   arr <- array(seq(4*3*4), c(4, 3, 4))/255
-  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1)
-  packed_set_dim(packed_ptr, 10L, 10L)
+  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1, as_ptr = TRUE)
+  packed_ptr_set_dim(packed_ptr, 10L, 10L)
   expect_true(TRUE)
 })
 
@@ -104,12 +104,12 @@ test_that("set_dim ok", {
 
 test_that("some error conditions", {
   arr <- array(seq(4*3*5), c(4, 3, 5))/255
-  expect_error(planar_to_packed(arr, format = 0L, maxval = 1), "3 or 4 planes")
+  expect_error(planar_to_packed(arr, format = 0L, maxval = 1, as_ptr = TRUE), "3 or 4 planes")
 
   expect_error(packed_to_planar(mtcars), "Can't unpack type")
 
   arr <- array(seq(4*3*4), c(4, 3, 4))/255
-  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1)
+  packed_ptr <- planar_to_packed(arr, format = 0L, maxval = 1, as_ptr = TRUE)
   expect_error(packed_to_planar(packed_ptr, format = 0L, nchannel = 8), 'nchannel')
 
 })