+++ NEWS.md
@@ -1,3 +1,9 @@
+# pixelweaver 0.1.2
+
+* Direct support for packed color stored in vectors of raw values
+* Internal refactor to reduce code duplication
+* Nomenclature tidy
+
 # pixelweaver 0.1.1
 
 * Make package a little more generic to support more pixel formats